# src/main/resources/META-INF/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
    ContactType:
      type: string
      enum:
      - Email
      - Telefoonnummer
      - ApplicatieId
    ContactgegevenRequest:
      description: Request object voor het toevoegen van een contactgegeven aan een
        partij
      type: object
      required:
      - identificatieType
      - identificatieNummer
      - type
      - waarde
      properties:
        identificatieType:
          $ref: "#/components/schemas/IdentificatieType"
        identificatieNummer:
          type: string
          pattern: ".*\\S.*"
        type:
          $ref: "#/components/schemas/ContactType"
        waarde:
          type: string
        scope:
          $ref: "#/components/schemas/ScopeRequest"
        teVerwijderenOp:
          $ref: "#/components/schemas/Instant"
    ContactgegevenResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        type:
          $ref: "#/components/schemas/ContactType"
        waarde:
          type: string
        isGeverifieerd:
          type: boolean
        isDefault:
          type: boolean
        createdAt:
          $ref: "#/components/schemas/Instant"
        lastUpdated:
          $ref: "#/components/schemas/Instant"
        teVerwijderenOp:
          $ref: "#/components/schemas/Instant"
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/ScopeResponse"
    ContactgegevenUpdateRequest:
      type: object
      required:
      - identificatieType
      - identificatieNummer
      - type
      - waarde
      properties:
        identificatieType:
          $ref: "#/components/schemas/IdentificatieType"
        identificatieNummer:
          type: string
          pattern: ".*\\S.*"
        type:
          $ref: "#/components/schemas/ContactType"
        waarde:
          type: string
        scope:
          $ref: "#/components/schemas/ScopeRequest"
        teVerwijderenOp:
          $ref: "#/components/schemas/Instant"
        id:
          $ref: "#/components/schemas/UUID"
        isDefault:
          type: boolean
...
    HttpProblem:
      description: HTTP Problem Response according to RFC9457 and RFC7807
      additionalProperties: true
      type: object
      properties:
        type:
          type: string
          format: uri
          description: A optional URI reference that identifies the problem type
          examples:
          - https://example.com/errors/not-found
        title:
          type: string
          description: "A optional, short, human-readable summary of the problem type"
          examples:
          - Not Found
        status:
          type: integer
          format: int32
          description: The HTTP status code for this occurrence of the problem
          examples:
          - 404
        detail:
          type: string
          description: A optional human-readable explanation specific to this occurrence
            of the problem
          examples:
          - Record not found
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of
            the problem
          examples:
          - https://api.example.com/errors/123
...
    IdentificatieType:
      type: string
      enum:
      - BSN
      - KVK
      - RSIN
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    PartijBulkRequest:
      type: object
      required:
      - identificaties
      properties:
        identificaties:
          type: array
          items:
            $ref: "#/components/schemas/PartijIdentificatieRequest"
          minItems: 1
    PartijIdentificatieRequest:
      type: object
      required:
      - identificatieType
      - identificatieNummer
      properties:
        identificatieType:
          $ref: "#/components/schemas/IdentificatieType"
        identificatieNummer:
          type: string
          pattern: ".*\\S.*"
    PartijRequest:
      description: Request object voor het ophalen van een Partij
      type: object
      required:
      - identificatieType
      - identificatieNummer
      properties:
        identificatieType:
          $ref: "#/components/schemas/IdentificatieType"
        identificatieNummer:
          type: string
          pattern: ".*\\S.*"
        dienstverlener:
          type: string
        dienstNaam:
          type: string
    PartijResponse:
      type: object
      properties:
        partijId:
          $ref: "#/components/schemas/UUID"
        identificaties:
          type: array
          items:
            $ref: "#/components/schemas/IdentificatieResponse"
        voorkeuren:
          type: array
          items:
            $ref: "#/components/schemas/VoorkeurResponse"
        contactgegevens:
          type: array
          items:
            $ref: "#/components/schemas/ContactgegevenResponse"
    ScopeRequest:
      type: object
      properties:
        dienstverlenerNaam:
          type: string
        dienstNaam:
          type: string
...
    TeVerwijderenOpRequest:
      type: object
      required:
      - id
      - identificatieType
      - identificatieNummer
      - dienstverlenerNaam
      - teVerwijderenOp
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        identificatieType:
          $ref: "#/components/schemas/IdentificatieType"
        identificatieNummer:
          type: string
          pattern: ".*\\S.*"
        dienstverlenerNaam:
          type: string
          pattern: ".*\\S.*"
        dienstNaam:
          type: string
        teVerwijderenOp:
          $ref: "#/components/schemas/Instant"
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
...
    VoorkeurRequest:
      type: object
      required:
      - identificatieType
      - identificatieNummer
      - voorkeurType
      - waarde
      properties:
        identificatieType:
          $ref: "#/components/schemas/IdentificatieType"
        identificatieNummer:
          type: string
          pattern: ".*\\S.*"
        voorkeurType:
          $ref: "#/components/schemas/VoorkeurType"
        waarde:
          type: string
        scope:
          $ref: "#/components/schemas/ScopeRequest"
        teVerwijderenOp:
          $ref: "#/components/schemas/Instant"
    VoorkeurResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        voorkeurType:
          $ref: "#/components/schemas/VoorkeurType"
        waarde:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        lastUpdated:
          $ref: "#/components/schemas/Instant"
        teVerwijderenOp:
          $ref: "#/components/schemas/Instant"
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/ScopeResponse"
...
    VoorkeurUpdateRequest:
      type: object
      required:
      - identificatieType
      - identificatieNummer
      - voorkeurType
      - waarde
      properties:
        identificatieType:
          $ref: "#/components/schemas/IdentificatieType"
        identificatieNummer:
          type: string
          pattern: ".*\\S.*"
        voorkeurType:
          $ref: "#/components/schemas/VoorkeurType"
        waarde:
          type: string
        scope:
          $ref: "#/components/schemas/ScopeRequest"
        teVerwijderenOp:
          $ref: "#/components/schemas/Instant"
        id:
          $ref: "#/components/schemas/UUID"
...
paths:
  /api/profielservice/v1/contactgegeven:
    put:
      summary: Update contactgegeven van een partij
      operationId: updateContactgegeven
      description: "Werk type, waarde en scope van een contactgegeven bij. Identificatie\
        \ kan niet aangepast worden."
      tags:
      - Profiel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactgegevenUpdateRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "200":
          description: Contactgegeven succesvol bijgewerkt
          content:
            application/json:
              schema: {}
        "400":
          description: Request body ontbreekt of is ongeldig
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "404":
          description: Contactgegeven of partij niet gevonden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
    post:
      summary: Toevoegen nieuwe contactgegeven voor een partij
      operationId: addContactgegeven
      description: Voegt een nieuwe contactgegeven toe. Creëert automatisch ontbrekende
        partijen.
      tags:
      - Profiel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactgegevenRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "201":
          description: Contactgegeven succesvol toegevoegd
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactgegevenResponse"
        "200":
          description: Contactgegeven was al geregistreerd voor deze partij en scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactgegevenResponse"
        "400":
          description: Request body ontbreekt of is ongeldig
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
  /api/profielservice/v1/contactgegeven/te-verwijderen-op:
    patch:
      summary: Stel te-verwijderen-op in voor een contactgegeven (Dienstverlener)
      operationId: updateContactgegevenTeVerwijderenOp
      description: Stelt of overschrijft de te-verwijderen-op datum voor een contactgegeven.
        Alleen toegestaan voor een Dienstverlener met een bestaande scope op het contactgegeven.
      tags:
      - Profiel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeVerwijderenOpRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "200":
          description: Te-verwijderen-op succesvol bijgewerkt
          content:
            application/json:
              schema: {}
        "400":
          description: Request body ontbreekt of is ongeldig
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "403":
          description: Dienstverlener heeft geen scope op dit contactgegeven
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "404":
          description: Contactgegeven of partij niet gevonden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
  /api/profielservice/v1/contactgegeven/{contactgegevenId}:
    delete:
      summary: Verwijder contactgegeven van een partij
      operationId: deleteContactgegeven
      description: Verwijdert een contactgegeven volledig
      tags:
      - Profiel
      parameters:
      - name: contactgegevenId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartijIdentificatieRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "204":
          description: Contactgegeven succesvol verwijderd
        "404":
          description: Contactgegeven of partij niet gevonden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
...
  /api/profielservice/v1/partij:
    post:
      summary: Ophalen profiel van een partij
      operationId: getPartij
      description: Haalt het profiel op van een partij
      tags:
      - Profiel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartijRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "200":
          description: Partij succesvol opgehaald
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartijResponse"
        "404":
          description: Partij niet gevonden of is verwijderd
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "400":
          description: Bad Request
  /api/profielservice/v1/partijen/bulk:
    post:
      summary: Ophalen profielen van meerdere partijen
      operationId: getPartijBulk
      description: Haalt profielen op van meerdere partijen. Niet-gevonden partijen
        worden stilzwijgend weggelaten.
      tags:
      - Profiel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartijBulkRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "200":
          description: Alle profielen succesvol opgehaald
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartijResponse"
        "206":
          description: Profielen gedeeltelijk opgehaald
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartijResponse"
        "400":
          description: Request body ontbreekt of is ongeldig
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "404":
          description: Geen enkel profiel gevonden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
  /api/profielservice/v1/voorkeur:
    put:
      summary: Update voorkeur van een partij
      operationId: updateVoorkeur
      description: "Werk type, waarde en scope van een voorkeur bij. Identificatie\
        \ kan niet aangepast worden."
      tags:
      - Profiel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoorkeurUpdateRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "200":
          description: Voorkeur succesvol bijgewerkt
          content:
            application/json:
              schema: {}
        "400":
          description: Request body ontbreekt of is ongeldig
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "404":
          description: Voorkeur of partij niet gevonden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
    post:
      summary: Toevoegen nieuwe voorkeur voor een partij
      operationId: addVoorkeur
      description: Voegt een nieuwe voorkeur toe. Creëert automatisch ontbrekende
        partijen.
      tags:
      - Profiel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoorkeurRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "201":
          description: Voorkeur succesvol toegevoegd
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoorkeurResponse"
        "200":
          description: Voorkeur was al geregistreerd voor deze partij en scope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoorkeurResponse"
        "400":
          description: Request body ontbreekt of is ongeldig
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
  /api/profielservice/v1/voorkeur/te-verwijderen-op:
    patch:
      summary: Stel te-verwijderen-op in voor een voorkeur (Dienstverlener)
      operationId: updateVoorkeurTeVerwijderenOp
      description: Stelt of overschrijft de te-verwijderen-op datum voor een voorkeur.
        Alleen toegestaan voor een Dienstverlener met een bestaande scope op de voorkeur.
      tags:
      - Profiel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeVerwijderenOpRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "200":
          description: Te-verwijderen-op succesvol bijgewerkt
          content:
            application/json:
              schema: {}
        "400":
          description: Request body ontbreekt of is ongeldig
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "403":
          description: Dienstverlener heeft geen scope op deze voorkeur
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "404":
          description: Voorkeur of partij niet gevonden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
  /api/profielservice/v1/voorkeur/{voorkeurId}:
    delete:
      summary: Verwijder voorkeur van een partij
      operationId: deleteVoorkeur
      description: Verwijdert een voorkeur volledig
      tags:
      - Profiel
      parameters:
      - name: voorkeurId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartijIdentificatieRequest"
        required: true
      responses:
        "500":
          description: Interne serverfout
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
        "204":
          description: Voorkeur succesvol verwijderd
        "404":
          description: Voorkeur of partij niet gevonden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/HttpProblem"
info:
  title: profiel-service API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value